+++ cdc/service_test.go
@@ -32,14 +32,15 @@ func Test_ServiceSingleEvent(t *testing.T) {
 	cl := &mockCluster{}
 	cl.leader.Store(true)
 
+	cfg := DefaultConfig()
+	cfg.Endpoint = testSrv.URL
+	cfg.MaxBatchSz = 1
+	cfg.MaxBatchDelay = 50 * time.Millisecond
 	svc := NewService(
+		cfg,
 		cl,
 		&mockStore{},
 		eventsCh,
-		testSrv.URL,
-		nil,                 // no TLS
-		1,                   // maxBatchSz – flush immediately
-		50*time.Millisecond, // maxBatchDelay – short for test
 	)
 	if err := svc.Start(); err != nil {
 		t.Fatalf("failed to start service: %v", err)
@@ -72,7 +73,7 @@ func Test_ServiceSingleEvent(t *testing.T) {
 			t.Fatalf("unexpected number of events in payload: %d", len(batch.Payload[0].Events))
 		}
 		if reflect.DeepEqual(batch.Payload[0].Events[0], evs.Events[0]) == false {
-			t.Fatalf("unexpected events in payload: %v", batch.Payload[0].Events)
+			t.Fatalf("unexpected events in payload: %v %v", batch.Payload[0].Events, evs.Events[0])
 		}
 	case <-time.After(2 * time.Second):
 		t.Fatalf("timeout waiting for HTTP POST")
@@ -86,6 +87,47 @@ func Test_ServiceSingleEvent(t *testing.T) {
 	cl.leader.Store(false)
 	eventsCh <- evs
 	pollExpvarUntil(t, numDroppedNotLeader, 1, 2*time.Second)
+}
+
+func Test_ServiceSingleEvent_LogOnly(t *testing.T) {
+	ResetStats()
+
+	// Channel for the service to receive events.
+	eventsCh := make(chan *proto.CDCEvents, 1)
+
+	cl := &mockCluster{}
+	cl.leader.Store(true)
+
+	cfg := DefaultConfig()
+	cfg.MaxBatchSz = 1
+	cfg.MaxBatchDelay = 50 * time.Millisecond
+	cfg.LogOnly = true
+	svc := NewService(
+		cfg,
+		cl,
+		&mockStore{},
+		eventsCh,
+	)
+	if err := svc.Start(); err != nil {
+		t.Fatalf("failed to start service: %v", err)
+	}
+	defer svc.Stop()
+
+	// Send one dummy event to the service.
+	ev := &proto.CDCEvent{
+		Op:       proto.CDCEvent_INSERT,
+		Table:    "foo",
+		NewRowId: 2,
+	}
+	evs := &proto.CDCEvents{
+		Index:  1,
+		Events: []*proto.CDCEvent{ev},
+	}
+	eventsCh <- evs
+
+	testPoll(t, func() bool {
+		return svc.HighWatermark() == evs.Index
+	}, 2*time.Second)
 }
 
 func Test_ServiceSingleEvent_Retry(t *testing.T) {
@@ -111,14 +153,15 @@ func Test_ServiceSingleEvent_Retry(t *testing.T) {
 	cl := &mockCluster{}
 	cl.leader.Store(true)
 
+	cfg := DefaultConfig()
+	cfg.Endpoint = testSrv.URL
+	cfg.MaxBatchSz = 1
+	cfg.MaxBatchDelay = 50 * time.Millisecond
 	svc := NewService(
+		cfg,
 		cl,
 		&mockStore{},
 		eventsCh,
-		testSrv.URL,
-		nil,                 // no TLS
-		1,                   // maxBatchSz – flush immediately
-		50*time.Millisecond, // maxBatchDelay – short for test
 	)
 	if err := svc.Start(); err != nil {
 		t.Fatalf("failed to start service: %v", err)
@@ -180,14 +223,15 @@ func Test_ServiceMultiEvent(t *testing.T) {
 	cl := &mockCluster{}
 	cl.leader.Store(true)
 
+	cfg := DefaultConfig()
+	cfg.Endpoint = testSrv.URL
+	cfg.MaxBatchSz = 2
+	cfg.MaxBatchDelay = time.Second
 	svc := NewService(
+		cfg,
 		cl,
 		&mockStore{},
 		eventsCh,
-		testSrv.URL,
-		nil, // no TLS
-		2,   // maxBatchSz – flush after 2 events
-		1*time.Second,
 	)
 	if err := svc.Start(); err != nil {
 		t.Fatalf("failed to start service: %v", err)
@@ -263,15 +307,15 @@ func Test_ServiceMultiEvent_Batch(t *testing.T) {
 	cl := &mockCluster{}
 	cl.leader.Store(true)
 
-	// Construct and start the service.
+	cfg := DefaultConfig()
+	cfg.Endpoint = testSrv.URL
+	cfg.MaxBatchSz = 2
+	cfg.MaxBatchDelay = 100 * time.Millisecond
 	svc := NewService(
+		cfg,
 		cl,
 		&mockStore{},
 		eventsCh,
-		testSrv.URL,
-		nil, // no TLS
-		2,   // maxBatchSz – flush after 2 events
-		100*time.Millisecond,
 	)
 	if err := svc.Start(); err != nil {
 		t.Fatalf("failed to start service: %v", err)