+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Implementation changes and bug fixes
 - [PR #2106](https://github.com/rqlite/rqlite/pull/2106), [PR #2126](https://github.com/rqlite/rqlite/pull/2126): Add initial CDC Service implementation.
 - [PR #2128](https://github.com/rqlite/rqlite/pull/2128): Support CDC highwatermarking control.
+- [PR #2129](https://github.com/rqlite/rqlite/pull/2129): CDC Service improvements and refactoring.
 
 ## v8.38.2 (June 23rd 2025)
 ### Implementation changes and bug fixes